move tests in integration-tests folder
